% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters.R
\name{backend_from_csem}
\alias{backend_from_csem}
\title{Create a backend skeleton for cSEM}
\usage{
backend_from_csem(estimator = NULL)
}
\arguments{
\item{estimator}{Optional function using \code{cSEM}.}
}
\value{
A \code{stresspls_backend}.
}
\description{
Create a backend skeleton for cSEM
}
\examples{
backend_from_csem()
}
