% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters.R
\name{backend_from_plspm}
\alias{backend_from_plspm}
\title{Create a backend skeleton for plspm}
\usage{
backend_from_plspm(estimator = NULL)
}
\arguments{
\item{estimator}{Optional function using \code{plspm}.}
}
\value{
A \code{stresspls_backend}.
}
\description{
Create a backend skeleton for plspm
}
\examples{
backend_from_plspm()
}
