% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters.R
\name{backend_from_seminr}
\alias{backend_from_seminr}
\title{Create a backend skeleton for seminr}
\usage{
backend_from_seminr(estimator = NULL)
}
\arguments{
\item{estimator}{Optional function using \code{seminr}.}
}
\value{
A \code{stresspls_backend}.
}
\description{
Create a backend skeleton for seminr
}
\examples{
backend_from_seminr()
}
