% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{calc_ci_width}
\alias{calc_ci_width}
\title{Calculate confidence interval width}
\usage{
calc_ci_width(x, type = c("paths", "weights"))
}
\arguments{
\item{x}{Numeric estimates or a \code{stresspls_fit_grid}.}

\item{type}{Estimate type for fit grids.}
}
\value{
A numeric value for numeric input, otherwise a tidy data frame.
}
\description{
Calculate confidence interval width
}
\examples{
calc_ci_width(c(0.1, 0.2, 0.4, 0.5))
}
