% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{calc_stability_index}
\alias{calc_stability_index}
\title{Calculate a stability index}
\usage{
calc_stability_index(x, type = c("paths", "weights"))
}
\arguments{
\item{x}{Numeric estimates or a \code{stresspls_fit_grid}.}

\item{type}{Estimate type for fit grids.}
}
\value{
A numeric value for numeric input, otherwise a tidy data frame.
}
\description{
Calculate a stability index
}
\examples{
calc_stability_index(c(0.1, 0.2, 0.3))
}
