% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneity.R
\name{compare_subgroup_weights}
\alias{compare_subgroup_weights}
\title{Compare subgroup weight estimates}
\usage{
compare_subgroup_weights(x)
}
\arguments{
\item{x}{A \code{stresspls_heterogeneity} object.}
}
\value{
A tidy comparison table.
}
\description{
Compare subgroup weight estimates
}
\examples{
# See `subgroup_heterogeneity()` for construction.
}
