% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-spec.R
\name{construct_names}
\alias{construct_names}
\title{Return construct names}
\usage{
construct_names(model)
}
\arguments{
\item{model}{A \code{stresspls_model_spec} object.}
}
\value{
A character vector of construct names.
}
\description{
\code{construct_names()} returns lower-order and higher-order construct names from
a model specification.
}
\examples{
image <- specify_construct("Image", c("img1", "img2"))
quality <- specify_construct("Quality", c("qual1", "qual2"))
brand_equity <- specify_hoc("BrandEquity", c("Image", "Quality"))
model <- specify_model(list(image, quality), hocs = list(brand_equity))
construct_names(model)
}
