% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{extract_prediction_metrics}
\alias{extract_prediction_metrics}
\title{Extract prediction metrics}
\usage{
extract_prediction_metrics(x)
}
\arguments{
\item{x}{A \code{stresspls_fit} or \code{stresspls_fit_grid}.}
}
\value{
A data frame of prediction metrics.
}
\description{
Extract prediction metrics
}
\examples{
fit <- as_stresspls_fit(list(prediction = data.frame(outcome = "Y",
                                                     metric = "rmse",
                                                     value = 1)))
extract_prediction_metrics(fit)
}
