% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{extract_r2}
\alias{extract_r2}
\title{Extract R-squared values}
\usage{
extract_r2(x)
}
\arguments{
\item{x}{A \code{stresspls_fit} or \code{stresspls_fit_grid}.}
}
\value{
A data frame of R-squared values.
}
\description{
Extract R-squared values
}
\examples{
fit <- as_stresspls_fit(list(r2 = data.frame(construct = "Y", r2 = 0.4)))
extract_r2(fit)
}
