% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{extract_weights}
\alias{extract_weights}
\title{Extract formative weight estimates}
\usage{
extract_weights(x)
}
\arguments{
\item{x}{A \code{stresspls_fit} or \code{stresspls_fit_grid}.}
}
\value{
A data frame of weight estimates.
}
\description{
Extract formative weight estimates
}
\examples{
fit <- as_stresspls_fit(list(weights = data.frame(construct = "A",
                                                  indicator = "a1",
                                                  estimate = 0.2)))
extract_weights(fit)
}
