% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables-plots-reports.R
\name{make_baseline_table}
\alias{make_baseline_table}
\title{Make a baseline result table}
\usage{
make_baseline_table(fit)
}
\arguments{
\item{fit}{A \code{stresspls_fit}.}
}
\value{
A data frame.
}
\description{
Make a baseline result table
}
\examples{
make_baseline_table(as_stresspls_fit(list()))
}
