% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables-plots-reports.R
\name{make_bootstrap_stability_table}
\alias{make_bootstrap_stability_table}
\title{Make a bootstrap stability table}
\usage{
make_bootstrap_stability_table(x)
}
\arguments{
\item{x}{A \code{stresspls_bootstrap}.}
}
\value{
A data frame.
}
\description{
Make a bootstrap stability table
}
\examples{
# See `bootstrap_stability()` for construction.
}
