% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables-plots-reports.R
\name{make_collinearity_table}
\alias{make_collinearity_table}
\title{Make a collinearity stress table}
\usage{
make_collinearity_table(x)
}
\arguments{
\item{x}{A \code{stresspls_collinearity_stress}.}
}
\value{
A data frame.
}
\description{
Make a collinearity stress table
}
\examples{
# See `collinearity_stress_test()` for construction.
}
