% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables-plots-reports.R
\name{make_heterogeneity_table}
\alias{make_heterogeneity_table}
\title{Make a heterogeneity table}
\usage{
make_heterogeneity_table(x)
}
\arguments{
\item{x}{A \code{stresspls_heterogeneity}.}
}
\value{
A data frame.
}
\description{
Make a heterogeneity table
}
\examples{
# See `subgroup_heterogeneity()` for construction.
}
