% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables-plots-reports.R
\name{make_indicator_perturbation_table}
\alias{make_indicator_perturbation_table}
\title{Make an indicator perturbation table}
\usage{
make_indicator_perturbation_table(x)
}
\arguments{
\item{x}{A \code{stresspls_fit_grid} or \code{stresspls_perturbation_grid}.}
}
\value{
A data frame.
}
\description{
Make an indicator perturbation table
}
\examples{
# See `perturb_indicators()` for construction.
}
