% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables-plots-reports.R
\name{make_simulation_design_table}
\alias{make_simulation_design_table}
\title{Make a simulation design table}
\usage{
make_simulation_design_table(x)
}
\arguments{
\item{x}{A \code{stresspls_simulation} object or design data frame.}
}
\value{
A data frame.
}
\description{
Make a simulation design table
}
\examples{
make_simulation_design_table(data.frame(n = 100))
}
