% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables-plots-reports.R
\name{make_simulation_results_table}
\alias{make_simulation_results_table}
\title{Make a simulation results table}
\usage{
make_simulation_results_table(x)
}
\arguments{
\item{x}{A \code{stresspls_simulation}.}
}
\value{
A data frame.
}
\description{
Make a simulation results table
}
\examples{
# See `run_simulation_grid()` for construction.
}
