% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{new_stresspls_result}
\alias{new_stresspls_result}
\title{Create a stressPLS result object}
\usage{
new_stresspls_result(data, model, grid, results, call = NULL, backend = NULL)
}
\arguments{
\item{data}{Original analysis data.}

\item{model}{Model specification.}

\item{grid}{A \code{stresspls_grid} object.}

\item{results}{A data frame of estimation results.}

\item{call}{The matched call.}

\item{backend}{Name of the backend or \code{NULL}.}
}
\value{
An object of class \code{stresspls_result}.
}
\description{
Create a stressPLS result object
}
\keyword{internal}
