% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{new_stresspls_summary}
\alias{new_stresspls_summary}
\title{Create a stressPLS summary object}
\usage{
new_stresspls_summary(result, status_counts)
}
\arguments{
\item{result}{A \code{stresspls_result} object.}

\item{status_counts}{A data frame of status counts.}
}
\value{
An object of class \code{stresspls_summary}.
}
\description{
Create a stressPLS summary object
}
\keyword{internal}
