% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables-plots-reports.R
\name{plot_indicator_stability_heatmap}
\alias{plot_indicator_stability_heatmap}
\title{Plot indicator stability as a heatmap}
\usage{
plot_indicator_stability_heatmap(x)
}
\arguments{
\item{x}{A \code{stresspls_fit_grid}.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot indicator stability as a heatmap
}
\examples{
# See `fit_perturbation_grid()` for construction.
}
