% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables-plots-reports.R
\name{plot_path_distribution}
\alias{plot_path_distribution}
\title{Plot path estimate distributions}
\usage{
plot_path_distribution(x)
}
\arguments{
\item{x}{A \code{stresspls_fit_grid} or data frame from \code{extract_paths()}.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot path estimate distributions
}
\examples{
plot_path_distribution(data.frame(from = "A", to = "B", estimate = 0.1))
}
