% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables-plots-reports.R
\name{plot_prediction_comparison}
\alias{plot_prediction_comparison}
\title{Plot prediction metric comparisons}
\usage{
plot_prediction_comparison(x)
}
\arguments{
\item{x}{A \code{stresspls_prediction_validation} or metric table.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot prediction metric comparisons
}
\examples{
plot_prediction_comparison(data.frame(outcome = "y", metric = "rmse", value = 1))
}
