% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables-plots-reports.R
\name{plot_simulation_results}
\alias{plot_simulation_results}
\title{Plot simulation results}
\usage{
plot_simulation_results(x)
}
\arguments{
\item{x}{A \code{stresspls_simulation}.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot simulation results
}
\examples{
# See `run_simulation_grid()` for construction.
}
