% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-stress.R
\name{plot_stress}
\alias{plot_stress}
\title{Plot stressPLS summaries}
\usage{
plot_stress(x, ...)
}
\arguments{
\item{x}{A \code{stresspls_result} or \code{stresspls_summary} object.}

\item{...}{Reserved for future methods.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot stressPLS summaries
}
\examples{
dat <- data.frame(x1 = 1:4, y = 2:5)
model <- list(indicators = "x1")
result <- stress_pls(dat, model)
plot_stress(result)
}
