% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables-plots-reports.R
\name{plot_stress_workflow}
\alias{plot_stress_workflow}
\title{Plot the stressPLS workflow}
\usage{
plot_stress_workflow()
}
\value{
A \code{ggplot} object.
}
\description{
Plot the stressPLS workflow
}
\examples{
plot_stress_workflow()
}
