% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables-plots-reports.R
\name{plot_vif_stress_curve}
\alias{plot_vif_stress_curve}
\title{Plot a VIF stress curve}
\usage{
plot_vif_stress_curve(x)
}
\arguments{
\item{x}{A \code{stresspls_collinearity_stress}.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot a VIF stress curve
}
\examples{
# See `collinearity_stress_test()` for construction.
}
