% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-spec.R
\name{required_indicators}
\alias{required_indicators}
\title{Return required manifest indicators}
\usage{
required_indicators(model)
}
\arguments{
\item{model}{A \code{stresspls_model_spec} object.}
}
\value{
A character vector of indicator names.
}
\description{
\code{required_indicators()} returns the unique manifest variables required by all
lower-order constructs in a model specification.
}
\examples{
image <- specify_construct("Image", c("img1", "img2"))
model <- specify_model(list(image))
required_indicators(model)
}
