% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{stressPLS-package}
\alias{stressPLS}
\alias{stressPLS-package}
\title{stressPLS: Stress-Test Formative Higher-Order Constructs in PLS-SEM}
\description{
Provides an initial framework for generating and reporting stress-test perturbation grids for formative higher-order constructs in partial least squares structural equation models. The package is designed around modular model specifications, perturbation generation, estimation backends, robustness scoring, and reporting.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://stresspls.github.io/stressPLS/}
  \item Report bugs at \url{https://github.com/stresspls/stressPLS/issues}
}

}
\author{
\strong{Maintainer}: stressPLS contributors \email{stresspls@example.com}

Authors:
\itemize{
  \item stressPLS contributors \email{stresspls@example.com}
}

}
\keyword{internal}
