% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collinearity.R
\name{summarise_collinearity_stress}
\alias{summarise_collinearity_stress}
\title{Summarise a collinearity stress test}
\usage{
summarise_collinearity_stress(x)
}
\arguments{
\item{x}{A \code{stresspls_collinearity_stress} object.}
}
\value{
A tidy data frame.
}
\description{
Summarise a collinearity stress test
}
\examples{
# See `collinearity_stress_test()` for construction.
}
