% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{summarise_simulation_results}
\alias{summarise_simulation_results}
\title{Summarise simulation results}
\usage{
summarise_simulation_results(x)
}
\arguments{
\item{x}{A \code{stresspls_simulation} object.}
}
\value{
A tidy data frame of simulation performance.
}
\description{
Summarise simulation results
}
\examples{
# See `run_simulation_grid()` for construction.
}
