% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise-stress.R
\name{summarise_stress}
\alias{summarise_stress}
\title{Summarise a stressPLS result}
\usage{
summarise_stress(x)
}
\arguments{
\item{x}{A \code{stresspls_result} or \code{stresspls_fit_grid} object.}
}
\value{
A \code{stresspls_summary} object.
}
\description{
Summarise a stressPLS result
}
\examples{
dat <- data.frame(x1 = 1:4, y = 2:5)
model <- list(indicators = "x1")
result <- stress_pls(dat, model)
summarise_stress(result)
}
